+++ .ci/set_dependency_version
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -e
+
 "$(dirname $0)"/../vendor/github.com/gardener/gardener/hack/.ci/set_dependency_version
 
 "$(dirname $0)"/../hack/.ci/set_dependency_version_dnsman
