+++ .dockerignore
@@ -7,6 +7,7 @@
 !docs/
 !example/
 !hack/
+!imagevector/
 !pkg/
 !test/
 !tools/
