+++ charts/charts.go
@@ -16,14 +16,10 @@ package charts
 
 import (
 	"embed"
-	_ "embed"
 )
 
-// Images YAML contains the contents of the images.yaml file.
-//go:embed images.yaml
-var ImagesYAML string
-
 // Internal contains the internal charts
+//
 //go:embed internal
 var Internal embed.FS
